Implementation of IR
Based on research into other compilers, it has become clear that having a SSA-based IR is incredibly helpful for analyzing/manipulating code and handling destructor insertion. For this reason, most of the code for compiling to LLVM IR has been cloned and tweaked to now write to a custom 'Re:C IR'.